security_platform/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.0
asyncpg==0.29.0
aioredis==2.0.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
plotly==5.18.0
flask==3.0.0
flask-socketio==5.3.4
python-socketio==5.10.0
prometheus-client==0.19.0
kubernetes==28.1.0
kopf==1.37.0
pyyaml==6.0.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
mypy==1.7.0
flake8==6.1.0
pre-commit==3.5.0
python-dotenv==1.0.0
PyGithub==2.1.1
jira==3.5.2
slack-sdk==3.21.3
redis==5.0.1
celery==5.3.4
aiohttp==3.9.0
